tests/requirements.txt
pytest<5.4
pytest-asyncio<0.11
pytest-cov>=2.8
pytest-twisted>=1.11
